% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sps_setup.R
\name{make_sps_setup}
\alias{make_sps_setup}
\title{Create an SPSS setup file (.sps) to use for reading in fixed-width text files}
\usage{
make_sps_setup(file_name, col_positions, col_names = NULL,
  col_labels = NULL, value_labels = NULL, missing_values = NULL)
}
\arguments{
\item{file_name}{Name of the file to be saved (e.g. "setup_file1"). There is no need to put
the .sps extention in the file name.}

\item{col_positions}{Either a vector of strings indicating the start and end position of each column
(e.g. "1-3", "4-5") or a vector of the widths of the columns (e.g. 3, 2).}

\item{col_names}{A vector of names for the columns. If none are provided, will automatically
create names based on column number (e.g. V1, V2, V3).}

\item{col_labels}{A vector of labels for the columns. These are often longer and more descriptive
than the col_names. These are the values used as column names if
real_names = TRUE in reading in the data.}

\item{value_labels}{A data.frame or tibble with the first column being the values and the second
column being the labels. The column name should go in the row above
the values and the column for that row should be an empty string.}

\item{missing_values}{A data.frame or tibble with the first column being the missing values and the second
column being the replacement. The column name should go in the row above
the values and the column for that row should be an empty string.}
}
\value{
Does not return any object. Saves the .sps file that is created.
}
\description{
make_sps_setup() is used to create the setup file used in reading in
fixed-width text files. Often the setup file comes with the data file but
in some cases (usually with government data) you will need to create
the setup file yourself.
}
\examples{
\dontrun{
value_labels <- data.frame(x = c("var1", 1:5),
y = c("",
      "label 1",
      "label 2",
      "label 3",
      "label 4",
      "label 5"))
make_sps_setup(file_name     = "example_name",
               col_positions = c(1, 3, 4, 2),
               col_names     = c("var1", "var2", "var3", "var4"),
               col_labels    = c("state name", "county",
                              "population", "census region code"),
               value_labels  = value_labels)
}
}
