% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{spss_reader}
\alias{spss_reader}
\title{Read ASCII file using SPSS Setup file}
\usage{
spss_reader(dataset_name, sps_name, value_label_fix = TRUE,
  real_names = TRUE)
}
\arguments{
\item{dataset_name}{Name of the ASCII file with the data}

\item{sps_name}{Name of the SPSS Setup file - should be a .sps or .txt file.}

\item{value_label_fix}{If TRUE, fixes value labels of the data. e.g. If a column is "sex" and has
values of 0 or 1, and the setup file says 0 = male and 1 = female, it will
make that change. The reader is much faster is this parameter is FALSE.}

\item{real_names}{If TRUE fixes column names from default column name in the SPSS setup file
(e.g. V1, V2) to the name is says the column is calle (e.g. age, sex, etc.)}
}
\value{
Data.frame of the data from the ASCII file
}
\description{
Read ASCII file using SPSS Setup file
}
\examples{

\dontrun{
example <- spss_reader(dataset_name = system.file("extdata",
"example_data.txt", package = "asciiReader"),
sps_name = system.file("extdata", "example_sps.txt", package = "asciiReader"))
}
# Does not fix value labels
example2 <- spss_reader(dataset_name = system.file("extdata",
"example_data.txt", package = "asciiReader"), sps_name = system.file("extdata",
 "example_sps.txt", package = "asciiReader"), value_label_fix = FALSE)

\dontrun{
# Keeps original column names
example3 <- spss_reader(dataset_name = system.file("extdata",
"example_data.txt", package = "asciiReader"), sps_name = system.file("extdata",
 "example_sps.txt", package = "asciiReader"), real_names = FALSE)
 }
}
