% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_spss_setup.R
\name{parse_spss}
\alias{parse_spss}
\title{Paese the SPSS setup file (.sps file)}
\usage{
parse_spss(sps_name)
}
\arguments{
\item{sps_name}{String of the name of the sps setup file.}
}
\description{
Paese the SPSS setup file (.sps file)
}
\examples{
sps_name <- system.file("extdata", "example_setup.sps",
  package = "asciiSetupReader")

\dontrun{
example  <- parse_spss(sps_name = sps_name)
}
}
