% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ascii_setup_addin.R
\name{read_ascii_setup_addin}
\alias{read_ascii_setup_addin}
\title{Launch an RStudio addin to select options for read_ascii_setup()}
\usage{
read_ascii_setup_addin()
}
\value{
the return of \code{\link[shiny]{runGadget}}
}
\description{
Launch an RStudio addin to select options for read_ascii_setup().
}
\examples{
\dontrun{
read_ascii_setup_addin()
}
}
